Implement Aria-label, Change migration
Added Aria-label so the accessibility reader app speaks what the buttons do (link, navigate to ... page)

migration changed so when using "php artisan migrate:fresh --seed" it doesn't respond with duplicate colum 'role'

+++ database/migrations/2026_05_28_112128_add_role_to_users_table.php
@@ -7,19 +7,24 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration
 {
     /**
-     * Run the migrations.
-     */
-    public function up()
-{
-    Schema::table('users', function (Blueprint $table) {
-        $table->string('role')->default('User');
-    });
-}
+    * Run the migrations.
+    */
 
-public function down()
-{
-    Schema::table('users', function (Blueprint $table) {
-        $table->dropColumn('role');
-    });
-}
+    public function up(): void
+    {
+        if (!Schema::hasColumn('users', 'role')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->string('role')->default('User');
+            });
+        }
+    }
+
+    public function down(): void
+    {
+        if (Schema::hasColumn('users', 'role')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('role');
+            });
+        }
+    }
 };
